+++ src/main/java/com/chownow/capstone/controllers/SearchExampleController.java
@@ -1,6 +1,6 @@
 package com.chownow.capstone.controllers;
 
-import com.chownow.capstone.dto.RecipeDto;
+import com.chownow.capstone.models.SpoonApiDto;
 import com.chownow.capstone.services.RecipeService;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,14 +27,14 @@ public class SearchExampleController {
     public String getRecipes(@RequestParam(required = false) String term,Model viewModel) throws InterruptedException, ParseException, IOException {
         viewModel.addAttribute("term", term);
         viewModel.addAttribute("spoonApi", spoonApi);
-        viewModel.addAttribute("recipe", new RecipeDto());
+        viewModel.addAttribute("recipe", new SpoonApiDto());
 
         return "recipes/search";
     }
 
     @PostMapping("/search")
     public String saveRecipes(Model viewModel,
-                              @ModelAttribute RecipeDto recipe){
+                              @ModelAttribute SpoonApiDto recipe){
 
         recipeService.saveRecipe(recipe);
